Struct rlottie::RGB

#[repr(C)]
pub struct RGB<ComponentType> {
    pub r: ComponentType,
    pub g: ComponentType,
    pub b: ComponentType,
}

The RGB pixel

The component type can be u8 (aliased as RGB8), u16 (aliased as RGB16), or any other type (but simple copyable types are recommended.)

Fields
r: ComponentType

Red

g: ComponentType

Green

b: ComponentType

Blue

Implementations

impl<T> RGB<T>

pub const fn new(r: T, g: T, b: T) -> RGB<T>

Convenience function for creating a new pixel The order of arguments is R,G,B

impl<T> RGB<T> where
T: Clone,

pub fn iter(&self) -> Cloned<Iter<'_, T>>

Iterate over color components (R, G, and B)

pub fn alpha(&self, a: T) -> RGBA<T, T>

Convenience function for converting to RGBA

pub fn new_alpha<A>(&self, a: A) -> RGBA<T, A>

Convenience function for converting to RGBA with alpha channel of a different type than type of the pixels