Struct rlottie::Animation

pub struct Animation(/* private fields */);

A lottie animation.

Implementations

impl Animation

pub fn from_file<P>(path: P) -> Option<Self>where P: AsRef<Path>,

Constructs an animation object from file path. This file needs to be in JSON format; if you want to read telegram’s tgs files, you need to decompress them first.

Note that the rlottie library might cache the file and/or its resources.

pub fn from_data<D, K, P>( json_data: D, cache_key: K, resource_path: P ) -> Option<Self>where D: Into<Vec<u8>>, K: Into<Vec<u8>>, P: AsRef<Path>,

Constructs an animation object from JSON string data. External resources are resolved relative to resource_path.

Note that the cache_key might be used by the rlottie library to cache the json data and/or its resources.

This method will panic if json_data or cache_key contain nul bytes.

pub fn size(&self) -> Size

Return the default viewport size of this animation.

pub fn duration(&self) -> f64

Return the total duration of this animation in seconds.

pub fn totalframe(&self) -> usize

Return the total number of frames in this animation.

pub fn framerate(&self) -> f64

Return the default framerate of this animation.

pub fn frame_at_pos(&self, pos: f32) -> usize

Maps position to frame number and returns it.

pub fn render(&mut self, frame_num: usize, surface: &mut Surface)

Render the contents of a frame onto the surface.

pub fn set_fill_color(&mut self, keypath: &str, color: Rgb)

pub fn set_fill_opacity(&mut self, keypath: &str, opacity: f64)

pub fn set_stroke_color(&mut self, keypath: &str, color: Rgb)

pub fn set_stroke_opacity(&mut self, keypath: &str, opacity: f64)

pub fn set_stroke_width(&mut self, keypath: &str, width: f64)

pub fn set_tr_position(&mut self, keypath: &str, x: f64, y: f64)

pub fn set_tr_scale(&mut self, keypath: &str, width: f64, height: f64)

pub fn set_tr_rotation(&mut self, keypath: &str, rotation: f64)